Install protobuf 3 manually for ARM platform

+++ runtime/examples/C/arraysum/arraysum-example.c
@@ -116,7 +116,8 @@ int main(int argc, char **argv) {
     init_array(arr, SZ);
 
     uint64_t golden = arraysum(arr);
-    printf("Golden output for run %d: %lu\n", run, golden);
+    printf("Golden output for run %d: %llu\n", run,
+           (long long unsigned int)golden);
 
     // Create argument list
     JobList *jl = tapasco_job_param_new();
@@ -151,7 +152,7 @@ int main(int argc, char **argv) {
 
     free(arr);
 
-    printf("FPGA output for run %d: %lu\n", run, r);
+    printf("FPGA output for run %d: %lu\n", run, (long long unsigned int)r);
     printf("\nRUN %d %s\n", run, r == golden ? "OK" : "NOT OK");
   }
